requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-jose
passlib[bcrypt]
python-dotenv
